+++ ui/src/app/index.tsx
@@ -3,7 +3,7 @@ import * as ReactDOM from 'react-dom';
 import App from './App';
 import './index.less';
 import {Provider} from 'react-redux';
-import {store} from '../store';
+import {store} from '../store/appState';
 
 ReactDOM.render(
     <Provider store={store}><App/></Provider>,
