+++ ui/src/console/screen/index.ts
@@ -1,3 +1,4 @@
 import Screen from './Screen';
 
 export default Screen;
+export {PixelState} from './Pixel';
