+++ ui/src/store/core/index.ts
@@ -1,3 +1,3 @@
 export {CoreActions} from './coreActions';
-export {CoreState} from './coreState';
+export {CoreState, DefaultSystemState} from './coreState';
 export {coreReducer} from './coreReducer';
