+++ ui/src/store/games/index.ts
@@ -1,0 +1,4 @@
+export {SnakeActions, SnakeAction} from './snakeActions';
+export {snakeEpic} from './snakeEpic';
+export {snakeGameReducer} from './snakeReducer';
+export {SnakeGameState, DefaultSnakeGameState} from './snakeState';
